+++ backend/src/main/java/shop/ninescent/mall/cartItem/repository/CartItemRepository.java
@@ -1,8 +1,8 @@
-package shop.ninescent.cartItem.repository;
+package shop.ninescent.mall.cartItem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import shop.ninescent.cartItem.domain.CartItem;
+import shop.ninescent.mall.cartItem.domain.CartItem;
 
 import java.util.List;
