+++ backend/src/main/java/shop/ninescent/mall/security/SecurityConfig.java
@@ -25,7 +25,7 @@ public class SecurityConfig {
         return http
                 .csrf(csrf -> csrf.disable())
                 .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/api/auth/**", "/api/public/**", "/api/review/**", "/api/qna/**").permitAll()
+                        .requestMatchers("/api/auth/**", "/api/public/**", "/api/**").permitAll()
                         .anyRequest().authenticated()
                 )
                 .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class)
